refactor: moved the basket class to common folder
