+++ src/components/cart/index.ts
@@ -1,18 +1,20 @@
 import Component from '../../common/component';
-import { IProduct, IPurchase } from '../../common/interface';
-import { products} from '../../products';
+import Basket from '../../common/basket';
+import { IPurchase, IProduct } from '../../common/interface';
 import './style-cart.scss';
 
 export default class Cart extends Component {
+  private basket: Basket;
 
   constructor(name: string) {
     super(name);
+    this.basket = new Basket();
   }
 
   public init(): void {
     console.log('cart');
     const btn = document.querySelector('.cart-summary__submit') as HTMLElement;
-    btn.addEventListener('click', () => window.basket.addProduct());
+    btn.addEventListener('click', () => this.basket.addProduct());
     this.draw();
     this.drawSummary();
     this.initEvents();
@@ -22,7 +24,7 @@ export default class Cart extends Component {
     const cartProducts = document.querySelector('.cart-products') as HTMLElement;
     cartProducts.innerHTML = '';
 
-    window.basket.purchases.forEach((el: IPurchase) => {
+    this.basket.purchases.forEach((el: IPurchase) => {
       const cartEl = this.createCartProduct(el.product);
       cartProducts.append(cartEl);
     });
@@ -34,7 +36,7 @@ export default class Cart extends Component {
     const productId = product?.dataset.id;
 
     if (productId) {
-      const newCount = window.basket.changeProductCount(productId, String(target.textContent));
+      const newCount = this.basket.changeProductCount(productId, String(target.textContent));
   
       if (newCount > 0) {
         const searchBlock = target.parentElement?.parentElement as HTMLElement;
@@ -43,8 +45,8 @@ export default class Cart extends Component {
   
         const searchTotal = target.parentElement?.parentElement?.parentElement as HTMLElement;
         const subTotal = searchTotal.querySelector('.cart-products__subtotal') as HTMLElement;
-        subTotal.textContent = `${(newCount * window.basket.getProduct(productId).price * 
-          (100 - window.basket.getProduct(productId).discountPercentage) / 100).toFixed(2)}$`;
+        subTotal.textContent = `${(newCount * this.basket.getProduct(productId).price * 
+          (100 - this.basket.getProduct(productId).discountPercentage) / 100).toFixed(2)}$`;
         this.drawSummary();
       } else {
         this.draw();
@@ -59,7 +61,7 @@ export default class Cart extends Component {
     const productId = product?.dataset.id;
 
     if (productId) {
-      window.basket.deleteProduct(productId);
+      this.basket.deleteProduct(productId);
       this.draw();
       this.drawSummary();
       this.initEvents();
@@ -68,13 +70,13 @@ export default class Cart extends Component {
 
   private drawSummary() {
     const countProducts = document.querySelector('.cart-summary__products-count') as HTMLElement;
-    countProducts.textContent = `${window.basket.getTotalCount()}`;
+    countProducts.textContent = `${this.basket.getTotalCount()}`;
 
     const countDiscount = document.querySelector('.cart-summary__discount-count') as HTMLElement;
-    countDiscount.textContent = `${window.basket.getTotalDiscount().toFixed(2)}$`;
+    countDiscount.textContent = `${this.basket.getTotalDiscount().toFixed(2)}$`;
 
     const countSumm = document.querySelector('.cart-summary__total-count') as HTMLElement;
-    countSumm.textContent = `${window.basket.getTotalSumm().toFixed(2)}$`;
+    countSumm.textContent = `${this.basket.getTotalSumm().toFixed(2)}$`;
   }
 
   private initEvents():void {
@@ -129,11 +131,11 @@ export default class Cart extends Component {
     liDiscount.textContent = `${product.discountPercentage}%`;
     img.src = product.thumbnail;
     img.alt = `${product.title}`;
-    count.textContent = `${window.basket.getProductCount(product.id)}`;
+    count.textContent = `${this.basket.getProductCount(product.id)}`;
     stock.textContent = `Stock: ${product.stock}`;
     btnPlus.textContent = '+';
     btnMinus.textContent = '-';
-    liSubtotal.textContent = `${(window.basket.getProductCount(product.id) * product.price * (100 - product.discountPercentage) / 100).toFixed(2)}$`;
+    liSubtotal.textContent = `${(this.basket.getProductCount(product.id) * product.price * (100 - product.discountPercentage) / 100).toFixed(2)}$`;
     
     div.append(btnPlus);
     div.append(btnMinus);
@@ -163,135 +165,5 @@ export default class Cart extends Component {
   }
 }
 
-export class Basket {
-  public purchases: IPurchase[];
-  public totalSumm: number;
-  public totalCount: number;
-  public totalDiscount: number;
 
-  constructor() {
-    const basketSave:string | null = localStorage.getItem('basket');
-    if (basketSave) {
-      const basketValue = JSON.parse(basketSave);
-      this.purchases = basketValue.purchases;
-      this.totalCount = basketValue.totalCount;
-      this.totalSumm = basketValue.totalSumm;
-      this.totalDiscount = basketValue.totalDiscount;
-    } else {
-      this.purchases = [];
-      this.totalCount = 0;
-      this.totalSumm = 0;
-      this.totalDiscount = 0;
-    }
-  }
-
-  public init(): void {
-    console.log(this.purchases);
-  }
-
-  public addProduct(): void {
-    const id: string | null = localStorage.getItem('productId');
-    if (id) {
-      if (this.isInBasket(id)) {
-        if (this.purchases[this.getPurchaseId(id)].count < this.purchases[this.getPurchaseId(id)].product.stock) {
-          this.purchases[this.getPurchaseId(id)].count++;
-        }
-      } else {
-        this.purchases.push({count: 1, product: this.getProduct(id)});
-        this.totalCount++;
-        this.totalSumm += this.getProduct(id).price;
-      }
-      this.setTotals();
-      this.setToLocalStorage();
-    }
-  }
-
-  public getProductCount(id: string): number {
-    let result = 0;
-    this.purchases.forEach((el: IPurchase) => {
-      if (el.product.id === id) result = el.count;
-    })
-    return result;
-  }
-
-  public changeProductCount(id: string, operation: string): number {
-    const purchaseId =  this.getPurchaseId(id);
-
-    if (operation === '+') {
-      if (this.purchases[purchaseId].count < this.purchases[purchaseId].product.stock) {
-        this.purchases[purchaseId].count += 1;
-      }
-    } else if (operation === '-') {
-      if (this.purchases[purchaseId].count > 1) {
-        this.purchases[purchaseId].count -= 1;
-      } else {
-        this.deleteProduct(id);
-        return 0;
-      }
-    }
-    this.setTotals();
-    this.setToLocalStorage();
-    return this.purchases[purchaseId].count;
-  }
-
-  private setTotals(): void {
-    let count = 0;
-    let summ = 0;
-    let discount = 0;
-    this.purchases.forEach((element: IPurchase) => {
-      count += element.count;
-      summ += element.count * element.product.price * (100 - element.product.discountPercentage) / 100;
-      discount += element.count * element.product.price * element.product.discountPercentage / 100;
-    })
-    this.totalCount = count;
-    this.totalSumm = summ;
-    this.totalDiscount = discount;
-  }
-
-  public getTotalSumm(): number {
-    return this.totalSumm;
-  }
-
-  public getTotalCount(): number {
-    return this.totalCount;
-  }
-
-  public getTotalDiscount(): number {
-    return this.totalDiscount;
-  }
-
-  public deleteProduct(id: string): void {
-    const purchaseId =  this.getPurchaseId(id);
-    this.purchases.splice(purchaseId, 1);
-    this.setTotals();
-    this.setToLocalStorage();
-  }
-
-  public getProduct(id: string | null): IProduct {
-    const productsArray: IProduct[] = products.filter((element) => element.id === id);
-    return JSON.parse(JSON.stringify(productsArray[0]));
-  }
-
-  public isInBasket(id: string): boolean {
-    let result = false;
-    this.purchases.forEach((el: IPurchase) => {
-      if (el.product.id === id) result = true;
-    })
-    return result;
-  }
-
-  private getPurchaseId(id: string): number {
-    let result = -1;
-    for (let i = 0; i < this.purchases.length; i++) {
-      if (this.purchases[i].product.id === id) {
-        result = i;
-      }
-    }
-    return result;
-  }
-
-  public setToLocalStorage(): void {
-    localStorage.setItem('basket', JSON.stringify(this));
-  }
-}
